memory allocations
need to check piazza questions and make sure my understanding of the
project is correct so that what i write moving forward is in line with
the project description.  my understanding should work if i continue
with it but should prob do what is actually correct if they've ACTUALLY
ANSWERED MY QUESTION YET

+++ Project_3/my_vm.h
@@ -2,6 +2,7 @@
 #define MY_VM_H_INCLUDED
 #include <stdbool.h>
 #include <stdlib.h>
+#include <stdio.h>
 
 //Assume the address space is 32 bits, so the max memory size is 4GB
 //Page size is 4KB
@@ -11,7 +12,7 @@
 #define PGSIZE 4096
 
 // Maximum size of your memory
-#define MAX_MEMSIZE 4*1024*1024*1024
+#define MAX_MEMSIZE (unsigned long long)4*1024*1024*1024
 
 #define MEMSIZE 1024*1024*1024
